레디스 캐시 설정 오류를 수정한다.

+++ api/src/main/java/com/dnd/sbooky/api/support/cache/dto/CacheResult.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 public class CacheResult<T> {
 
     private final T data;
-    private final Integer delta;
+    private final Long delta;
     private final Long remainingTtl;
     private final boolean cacheHit;
